+++ ios/Extension/UIImage.swift
@@ -1,0 +1,29 @@
+//
+//  UIColor.swift
+//  CocoaAsyncSocket
+//
+//  Created by BẢO HÀ on 12/09/2023.
+//
+
+import UIKit
+
+extension UIImage {
+    func getTintColor(_ color: UIColor) -> UIImage? {
+        if #available(iOS 13.0, *) {
+            return self.withTintColor(color, renderingMode: .alwaysOriginal)
+        } else {
+            UIGraphicsBeginImageContextWithOptions(size, false, scale)
+            // 1
+            let drawRect = CGRect(x: 0, y: 0, width: size.width, height: size.height)
+            // 2
+            color.setFill()
+            UIRectFill(drawRect)
+            // 3
+            draw(in: drawRect, blendMode: .destinationIn, alpha: 1)
+
+            let tintedImage = UIGraphicsGetImageFromCurrentImageContext()
+            UIGraphicsEndImageContext()
+            return tintedImage!
+        }
+    }
+}
