+++ ios/MultipleImagePicker.swift
@@ -336,27 +336,3 @@ extension UIViewController {
         return topMostViewController
     }
 }
-
-extension UIColor {
-    convenience init(hex: String) {
-        var cString: String = hex.trimmingCharacters(in: .whitespacesAndNewlines).uppercased()
-        
-        if cString.hasPrefix("#") {
-            cString.remove(at: cString.startIndex)
-        }
-        
-        if (cString.count) != 6 {
-            self.init(red: 0.5, green: 0.5, blue: 0.5, alpha: 1.0) // Màu mặc định nếu chuỗi không hợp lệ
-        } else {
-            var rgbValue: UInt64 = 0
-            Scanner(string: cString).scanHexInt64(&rgbValue)
-            
-            self.init(
-                red: CGFloat((rgbValue & 0xFF0000) >> 16) / 255.0,
-                green: CGFloat((rgbValue & 0x00FF00) >> 8) / 255.0,
-                blue: CGFloat(rgbValue & 0x0000FF) / 255.0,
-                alpha: CGFloat(1.0)
-            )
-        }
-    }
-}
